+++ features/fixture_resources/lib/scenarios/scenario.dart
@@ -1,4 +1,5 @@
 // ignore_for_file: avoid_print
+import 'package:bugsnag_flutter/bugsnag_flutter.dart';
 import 'package:bugsnag_flutter_performance/bugsnag_flutter_performance.dart';
 import 'package:flutter/widgets.dart';
 
@@ -7,6 +8,7 @@ import '../main.dart';
 
 abstract class Scenario {
   String? extraConfig;
+  BugsnagEndpointConfiguration? endpointConfiguration;
   void Function()? runCommandCallback;
 
   Future<void> clearPersistentData() async {
@@ -35,10 +37,16 @@ abstract class Scenario {
     bugsnag_performance.setExtraConfig("instrumentNavigation", value);
   }
 
+  void setInstrumentsViewLoad(bool value) {
+    bugsnag_performance.setExtraConfig("instrumentViewLoad", value);
+  }
+
   Future<void> startBugsnag({
     String? releaseStage,
     List<String>? enabledReleaseStages,
+    List<RegExp>? tracePropagationUrls,
     String? appVersion,
+    bool shouldUseNotifier = false,
   }) async {
     bugsnag_performance.setExtraConfig("instrumentAppStart", false);
     bugsnag_performance.setExtraConfig("probabilityValueExpireTime", 1000);
@@ -47,8 +55,16 @@ abstract class Scenario {
       endpoint: Uri.parse('${FixtureConfig.MAZE_HOST}/traces'),
       releaseStage: releaseStage,
       enabledReleaseStages: enabledReleaseStages,
+      tracePropagationUrls: tracePropagationUrls,
       appVersion: appVersion,
     );
+    if (shouldUseNotifier && endpointConfiguration != null) {
+      await bugsnag.start(
+        apiKey: '12312312312312312312312312312312',
+        endpoints: endpointConfiguration!,
+        autoDetectErrors: false,
+      );
+    }
   }
 
   void invokeMethod(String name) {}