+++ packages/bugsnag_flutter_performance/lib/src/uploader/sampler.dart
@@ -10,7 +10,7 @@ abstract class Sampler {
   Future<bool> sample(BugsnagPerformanceSpan span);
   Future<List<BugsnagPerformanceSpan>> sampled(
       List<BugsnagPerformanceSpan> spans);
-  void handleResponseHeaders(HttpHeaders headers);
+  Future<void> handleResponseHeaders(HttpHeaders headers);
 }
 
 class SamplerImpl implements Sampler {
@@ -64,11 +64,12 @@ class SamplerImpl implements Sampler {
       (span as BugsnagPerformanceSpanImpl?)
           ?.updateSamplingProbability(samplingProbability);
     }
+    (span as BugsnagPerformanceSpanImpl?)?.isSampled = isSampled;
     return isSampled;
   }
 
   @override
-  void handleResponseHeaders(HttpHeaders headers) async {
+  Future<void> handleResponseHeaders(HttpHeaders headers) async {
     final samplingProbabilityHeader =
         headers.value('Bugsnag-Sampling-Probability');
     if (samplingProbabilityHeader == null) {